fix: remove trailing blank line from block scalars ending with newline
When a string ends with \n and is serialized as a block scalar (|),
split('\n') produces an empty trailing element. Previously this caused
an extra blank line to be emitted. Now we skip the trailing empty
element since the newline is implicit in YAML block scalars.

This improves compatibility with Go's yaml.v3 output.